backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyGithub==2.1.1
python-pptx==0.6.23
markdown2==2.4.9
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0